// lib/bigcommerce/schema.d.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  /** The `BigDecimal` scalar type represents signed fractional values with arbitrary precision. */
  BigDecimal: any;
  /** The `Long` scalar type represents non-fractional signed whole numeric values. Long can represent values between -(2^63) and 2^63 - 1. */
  Long: any;
};
// ...
export type Brand = Node & {
  __typename?: 'Brand';
  /** The ID of an object */
  id: Scalars['ID'];
  /** Id of the brand. */
  entityId: Scalars['Int'];
  /** Name of the brand. */
  name: Scalars['String'];
  /** Default image for brand. */
  defaultImage?: Maybe<Image>;
  /** Page title for the brand. */
  pageTitle: Scalars['String'];
  /** Meta description for the brand. */
  metaDesc: Scalars['String'];
  /** Meta keywords for the brand. */
  metaKeywords: Array<Scalars['String']>;
  /** Search keywords for the brand. */
  searchKeywords: Array<Scalars['String']>;
  /** Path for the brand page. */
  path: Scalars['String'];
  products: ProductConnection;
  /** Metafield data related to a brand. */
  metafields: MetafieldConnection;
};
// ...
export type BulkPricingTier = {
  /** Minimum item quantity that applies to this bulk pricing tier. */
  minimumQuantity: Scalars['Int'];
  /** Maximum item quantity that applies to this bulk pricing tier - if not defined then the tier does not have an upper bound. */
  maximumQuantity?: Maybe<Scalars['Int']>;
};
// ...
export type CategoryConnection = {
  __typename?: 'CategoryConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<CategoryEdge>>>;
};

/** An edge in a connection. */
export type CategoryEdge = {
  __typename?: 'CategoryEdge';
  /** The item at the end of the edge. */
  node: Category;
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
};
// ...
export type CustomFieldConnection = {
  __typename?: 'CustomFieldConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<CustomFieldEdge>>>;
};

/** An edge in a connection. */
export type CustomFieldEdge = {
  __typename?: 'CustomFieldEdge';
  /** The item at the end of the edge. */
  node: CustomField;
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
};
// ...
export type DateTimeExtended = {
  __typename?: 'DateTimeExtended';
  /** ISO-8601 formatted date in UTC */
  utc: Scalars['DateTime'];
};
// ...
export type Image = {
  __typename?: 'Image';
  /** Absolute path to image using store CDN. */
  url: Scalars['String'];
  /** Absolute path to original image using store CDN. */
  urlOriginal: Scalars['String'];
  /** Text description of an image that can be used for SEO and/or accessibility purposes. */
  altText: Scalars['String'];
  /** Indicates whether this is the primary image. */
  isDefault: Scalars['Boolean'];
};

/** Image */
export type ImageUrlArgs = {
  width: Scalars['Int'];
  height?: Maybe<Scalars['Int']>;
};

/** A connection to a list of items. */
export type ImageConnection = {
  __typename?: 'ImageConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<ImageEdge>>>;
};

/** An edge in a connection. */
export type ImageEdge = {
  __typename?: 'ImageEdge';
  /** The item at the end of the edge. */
  node: Image;
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
};
// ...
export type Measurement = {
  __typename?: 'Measurement';
  /** Unformatted weight measurement value. */
  value: Scalars['Float'];
  /** Unit of measurement. */
  unit: Scalars['String'];
};

/** A connection to a list of items. */
export type MetafieldConnection = {
  __typename?: 'MetafieldConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<MetafieldEdge>>>;
};
// ...
export type Money = {
  __typename?: 'Money';
  /** Currency code of the current money. */
  currencyCode: Scalars['String'];
  /** The amount of money. */
  value: Scalars['BigDecimal'];
};

/** A min and max pair of money objects */
export type MoneyRange = {
  __typename?: 'MoneyRange';
  /** Minimum money object. */
  min: Money;
  /** Maximum money object. */
  max: Money;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID'];
};
// ...
export type OptionConnection = {
  __typename?: 'OptionConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<OptionEdge>>>;
};

/** An edge in a connection. */
export type OptionEdge = {
  __typename?: 'OptionEdge';
  /** The item at the end of the edge. */
  node: ProductOption;
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']>;
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']>;
};

/** The min and max range of prices that apply to this product. */
export type PriceRanges = {
  __typename?: 'PriceRanges';
  /** Product price min/max range. */
  priceRange: MoneyRange;
  /** Product retail price min/max range. */
  retailPriceRange?: Maybe<MoneyRange>;
};

/** The various prices that can be set on a product. */
export type Prices = {
  __typename?: 'Prices';
  /** Calculated price of the product. */
  price: Money;
  /** Sale price of the product. */
  salePrice?: Maybe<Money>;
  /** Original price of the product. */
  basePrice?: Maybe<Money>;
  /** Retail price of the product. */
  retailPrice?: Maybe<Money>;
  /** Minimum advertised price of the product. */
  mapPrice?: Maybe<Money>;
  /** Product price min/max range. */
  priceRange: MoneyRange;
  /** Product retail price min/max range. */
  retailPriceRange?: Maybe<MoneyRange>;
  /** The difference between the retail price (MSRP) and the current price, which can be presented to the shopper as their savings. */
  saved?: Maybe<Money>;
  /** List of bulk pricing tiers applicable to a product or variant. */
  bulkPricing: Array<BulkPricingTier>;
};

/** Product */
export type Product = Node & {
  __typename?: 'Product';
  /** The ID of an object */
  id: Scalars['ID'];
  /** Id of the product. */
  entityId: Scalars['Int'];
  /** Default product variant when no options are selected. */
  sku: Scalars['String'];
  /** Relative URL path to product page. */
  path: Scalars['String'];
  /** Name of the product. */
  name: Scalars['String'];
  /** Description of the product. */
  description: Scalars['String'];
  /** Description of the product in plain text. */
  plainTextDescription: Scalars['String'];
  /** Warranty information of the product. */
  warranty: Scalars['String'];
  /** Minimum purchasable quantity for this product in a single order. */
  minPurchaseQuantity?: Maybe<Scalars['Int']>;
  /** Maximum purchasable quantity for this product in a single order. */
  maxPurchaseQuantity?: Maybe<Scalars['Int']>;
  /** Absolute URL path for adding a product to cart. */
  addToCartUrl: Scalars['String'];
  /** Absolute URL path for adding a product to customer's wishlist. */
  addToWishlistUrl: Scalars['String'];
  /** Prices object determined by supplied product ID, variant ID, and selected option IDs. */
  prices?: Maybe<Prices>;
  /**
   * The minimum and maximum price of this product based on variant pricing and/or modifier price rules.
   * @deprecated Use priceRanges inside prices node instead.
   */
  priceRanges?: Maybe<PriceRanges>;
  /** Weight of the product. */
  weight?: Maybe<Measurement>;
  /** Height of the product. */
  height?: Maybe<Measurement>;
  /** Width of the product. */
  width?: Maybe<Measurement>;
  /** Depth of the product. */
  depth?: Maybe<Measurement>;
  /** Product options. */
  options: OptionConnection;
  /** Product options. */
  productOptions: ProductOptionConnection;
  /** Summary of the product reviews, includes the total number of reviews submitted and summation of the ratings on the reviews (ratings range from 0-5 per review). */
  reviewSummary: Reviews;
  /** Type of product, ex: physical, digital */
  type: Scalars['String'];
  /**
   * The availability state of the product.
   * @deprecated Use status inside availabilityV2 instead.
   */
  availability: Scalars['String'];
  /**
   * A few words telling the customer how long it will normally take to ship this product, such as 'Usually ships in 24 hours'.
   * @deprecated Use description inside availabilityV2 instead.
   */
  availabilityDescription: Scalars['String'];
  /** The availability state of the product. */
  availabilityV2: ProductAvailability;
  /** List of categories associated with the product. */
  categories: CategoryConnection;
  /** Brand associated with the product. */
  brand?: Maybe<Brand>;
  /** Variants associated with the product. */
  variants: VariantConnection;
  /** Custom fields of the product. */
  customFields: CustomFieldConnection;
  /** A list of the images for a product. */
  images: ImageConnection;
  /** Default image for a product. */
  defaultImage?: Maybe<Image>;
  /** Related products for this product. */
  relatedProducts: RelatedProductsConnection;
  /** Inventory information of the product. */
  inventory: ProductInventory;
  /** Metafield data related to a product. */
  metafields: MetafieldConnection;
  /**
   * Product creation date
   * @deprecated Alpha version. Do not use in production.
   */
  createdAt: DateTimeExtended;
};
// ...
export type ProductAvailability = {
  /** The availability state of the product. */
  status: ProductAvailabilityStatus;
  /** A few words telling the customer how long it will normally take to ship this product, such as 'Usually ships in 24 hours'. */
  description: Scalars['String'];
};

/** Product availability status */
export enum ProductAvailabilityStatus {
  Available = 'Available',
  Preorder = 'Preorder',
  Unavailable = 'Unavailable'
}
// ...
export type ProductConnection = {
  __typename?: 'ProductConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<ProductEdge>>>;
};

/** An edge in a connection. */
export type ProductEdge = {
  __typename?: 'ProductEdge';
  /** The item at the end of the edge. */
  node: Product;
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
};

/** Product Inventory Information */
export type ProductInventory = {
  __typename?: 'ProductInventory';
  /** Indicates whether this product is in stock. */
  isInStock: Scalars['Boolean'];
  /** Indicates whether this product's inventory is being tracked on variant level. If true, you may wish to check the variants node to understand the true inventory of each individual variant, rather than relying on this product-level aggregate to understand how many items may be added to cart. */
  hasVariantInventory: Scalars['Boolean'];
  /** Aggregated product inventory information. This data may not be available if not set or if the store's Inventory Settings have disabled displaying stock levels on the storefront. */
  aggregated?: Maybe<AggregatedInventory>;
};
// ...
export type ProductOptionConnection = {
  __typename?: 'ProductOptionConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<ProductOptionEdge>>>;
};

/** An edge in a connection. */
export type ProductOptionEdge = {
  __typename?: 'ProductOptionEdge';
  /** The item at the end of the edge. */
  node: CatalogProductOption;
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
};
// ...
export type RelatedProductsConnection = {
  __typename?: 'RelatedProductsConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<RelatedProductsEdge>>>;
};

/** An edge in a connection. */
export type RelatedProductsEdge = {
  __typename?: 'RelatedProductsEdge';
  /** The item at the end of the edge. */
  node: Product;
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
};

/** Review Rating Summary */
export type Reviews = {
  __typename?: 'Reviews';
  /** Total number of reviews on product. */
  numberOfReviews: Scalars['Int'];
  /** Summation of rating scores from each review. */
  summationOfRatings: Scalars['Int'];
};
// ...
export type VariantConnection = {
  __typename?: 'VariantConnection';
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<VariantEdge>>>;
};

/** An edge in a connection. */
export type VariantEdge = {
  __typename?: 'VariantEdge';
  /** The item at the end of the edge. */
  node: Variant;
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
};